# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Olake is fastest open-source tool for replicating Databases to Apache Iceberg or Data Lakehouse.",
        "title": "Olake UI API",
        "contact": {
            "name": "OLake",
            "url": "https://olake.io/contact/",
            "email": "hello@olake.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/auth/check": {
            "get": {
                "description": "Verify if the current user session is active and valid.",
                "tags": [
                    "Authentication"
                ],
                "summary": "Check authentication status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "Authenticate a user and create a new session.",
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "413": {
                        "description": "payload too large",
                        "schema": {
                            "$ref": "#/definitions/dto.Error413Response"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "Register a new user account with the provided details.",
                "tags": [
                    "Authentication"
                ],
                "summary": "User signup",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "409": {
                        "description": "user already exists",
                        "schema": {
                            "$ref": "#/definitions/dto.Error409Response"
                        }
                    },
                    "413": {
                        "description": "payload too large",
                        "schema": {
                            "$ref": "#/definitions/dto.Error413Response"
                        }
                    },
                    "500": {
                        "description": "failed to create user",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
        "/telemetry-id": {
            "get": {
                "description": "Retrieve the unique telemetry identifier and current UI version.",
                "tags": [
                    "Internal"
                ],
                "summary": "Get telemetry ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TelemetryIDResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "dto.Error400Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Bad request"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.Error401Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Authentication required"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.Error409Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Resource already exists"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.Error413Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Payload too large"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.Error500Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal server error"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "dto.TelemetryIDResponse": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string",
                    "example": "1234567890abcdef1234567890abcdef"
                },
                "version": {
                    "type": "string",
                    "example": "v0.2.5"
                }
            }
        },
...
    },
...
}